requirements.txt
PyGithub
loguru
